# pkg/apiserver/docs/swagger.yaml
definitions:
...
  api.Pagination:
    properties:
      page:
        description: 当前页索引，从 1 开始，为 0 时，会自动重置为默认值 constants.DefaultPage
        type: integer
      pageSize:
        description: |-
          每页数据量，为 -1 时表示查询全部，为 0 时会重置为默认值
          constants.DefaultPageSize
        type: integer
      pages:
        description: 总页数
        type: integer
      total:
        description: 总共有多少条目，请求时可以不用传递
        type: integer
    type: object
  api.RspFailBody:
    properties:
      description:
        type: string
      errcode:
        type: integer
    type: object
...
  api.Volume:
    properties:
      apiVersion:
        description: |-
          APIVersion defines the versioned schema of this representation of an object.
          Servers should convert recognized schemas to the latest internal value, and
          may reject unrecognized values.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources
          +optional
        type: string
      kind:
        description: |-
          Kind is a string value representing the REST resource this object represents.
          Servers may infer this from the endpoint the client submits requests to.
          Cannot be updated.
          In CamelCase.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds
          +optional
        type: string
      metadata:
        $ref: '#/definitions/v1.ObjectMeta'
      spec:
        $ref: '#/definitions/v1alpha1.LocalVolumeSpec'
      status:
        $ref: '#/definitions/v1alpha1.LocalVolumeStatus'
    type: object
...
  api.VolumeConvertOperation:
    properties:
      apiVersion:
        description: |-
          APIVersion defines the versioned schema of this representation of an object.
          Servers should convert recognized schemas to the latest internal value, and
          may reject unrecognized values.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources
          +optional
        type: string
      kind:
        description: |-
          Kind is a string value representing the REST resource this object represents.
          Servers may infer this from the endpoint the client submits requests to.
          Cannot be updated.
          In CamelCase.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds
          +optional
        type: string
      metadata:
        $ref: '#/definitions/v1.ObjectMeta'
      spec:
        $ref: '#/definitions/v1alpha1.LocalVolumeConvertSpec'
      status:
        $ref: '#/definitions/v1alpha1.LocalVolumeConvertStatus'
    type: object
  api.VolumeConvertReqBody:
    properties:
      abort:
        type: boolean
    type: object
  api.VolumeConvertRspBody:
    properties:
      data:
        $ref: '#/definitions/api.VolumeConvertInfo'
    type: object
...
  api.VolumeList:
    properties:
      items:
        items:
          $ref: '#/definitions/api.Volume'
        type: array
      pagination:
        allOf:
        - $ref: '#/definitions/api.Pagination'
        description: page 信息
    type: object
...
  api.VolumeMigrateOperation:
    properties:
      apiVersion:
        description: |-
          APIVersion defines the versioned schema of this representation of an object.
          Servers should convert recognized schemas to the latest internal value, and
          may reject unrecognized values.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources
          +optional
        type: string
      kind:
        description: |-
          Kind is a string value representing the REST resource this object represents.
          Servers may infer this from the endpoint the client submits requests to.
          Cannot be updated.
          In CamelCase.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds
          +optional
        type: string
      metadata:
        $ref: '#/definitions/v1.ObjectMeta'
      spec:
        $ref: '#/definitions/v1alpha1.LocalVolumeMigrateSpec'
      status:
        $ref: '#/definitions/v1alpha1.LocalVolumeMigrateStatus'
    type: object
  api.VolumeMigrateReqBody:
    properties:
      abort:
        type: boolean
      selectedNode:
        type: string
      srcNode:
        type: string
    type: object
  api.VolumeMigrateRspBody:
    properties:
      data:
        $ref: '#/definitions/api.VolumeMigrateInfo'
    type: object
  api.VolumeOperationByVolume:
    properties:
      VolumeConvertOperations:
        description: VolumeConvertOperations
        items:
          $ref: '#/definitions/api.VolumeConvertOperation'
        type: array
      VolumeExpandOperations:
        description: VolumeExpandOperations
        items:
          $ref: '#/definitions/api.VolumeExpandOperation'
        type: array
      volumeMigrateOperations:
        description: |-
          // OperationList
          OperationList []Operation `json:"items"`
          VolumeMigrateOperations
        items:
          $ref: '#/definitions/api.VolumeMigrateOperation'
        type: array
      volumeName:
        description: VolumeName
        type: string
    type: object
...
  api.VolumeReplicaList:
    properties:
      volumeName:
        description: volume name
        type: string
      volumeReplicas:
        description: VolumeReplicas
        items:
          $ref: '#/definitions/api.VolumeReplica'
        type: array
    type: object
...
  v1.ObjectMeta:
    properties:
      annotations:
        additionalProperties:
          type: string
        description: |-
          Annotations is an unstructured key value map stored with a resource that may be
          set by external tools to store and retrieve arbitrary metadata. They are not
          queryable and should be preserved when modifying objects.
          More info: http://kubernetes.io/docs/user-guide/annotations
          +optional
        type: object
      clusterName:
        description: |-
          The name of the cluster which the object belongs to.
          This is used to distinguish resources with same name and namespace in different clusters.
          This field is not set anywhere right now and apiserver is going to ignore it if set in create or update request.
          +optional
        type: string
      creationTimestamp:
        description: |-
          CreationTimestamp is a timestamp representing the server time when this object was
          created. It is not guaranteed to be set in happens-before order across separate operations.
          Clients may not set this value. It is represented in RFC3339 form and is in UTC.

          Populated by the system.
          Read-only.
          Null for lists.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#metadata
          +optional
        type: string
      deletionGracePeriodSeconds:
        description: |-
          Number of seconds allowed for this object to gracefully terminate before
          it will be removed from the system. Only set when deletionTimestamp is also set.
          May only be shortened.
          Read-only.
          +optional
        type: integer
      deletionTimestamp:
        description: |-
          DeletionTimestamp is RFC 3339 date and time at which this resource will be deleted. This
          field is set by the server when a graceful deletion is requested by the user, and is not
          directly settable by a client. The resource is expected to be deleted (no longer visible
          from resource lists, and not reachable by name) after the time in this field, once the
          finalizers list is empty. As long as the finalizers list contains items, deletion is blocked.
          Once the deletionTimestamp is set, this value may not be unset or be set further into the
          future, although it may be shortened or the resource may be deleted prior to this time.
          For example, a user may request that a pod is deleted in 30 seconds. The Kubelet will react
          by sending a graceful termination signal to the containers in the pod. After that 30 seconds,
          the Kubelet will send a hard termination signal (SIGKILL) to the container and after cleanup,
          remove the pod from the API. In the presence of network partitions, this object may still
          exist after this timestamp, until an administrator or automated process can determine the
          resource is fully terminated.
          If not set, graceful deletion of the object has not been requested.

          Populated by the system when a graceful deletion is requested.
          Read-only.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#metadata
          +optional
        type: string
      finalizers:
        description: |-
          Must be empty before the object is deleted from the registry. Each entry
          is an identifier for the responsible component that will remove the entry
          from the list. If the deletionTimestamp of the object is non-nil, entries
          in this list can only be removed.
          Finalizers may be processed and removed in any order.  Order is NOT enforced
          because it introduces significant risk of stuck finalizers.
          finalizers is a shared field, any actor with permission can reorder it.
          If the finalizer list is processed in order, then this can lead to a situation
          in which the component responsible for the first finalizer in the list is
          waiting for a signal (field value, external system, or other) produced by a
          component responsible for a finalizer later in the list, resulting in a deadlock.
          Without enforced ordering finalizers are free to order amongst themselves and
          are not vulnerable to ordering changes in the list.
          +optional
          +patchStrategy=merge
        items:
          type: string
        type: array
      generateName:
        description: |-
          GenerateName is an optional prefix, used by the server, to generate a unique
          name ONLY IF the Name field has not been provided.
          If this field is used, the name returned to the client will be different
          than the name passed. This value will also be combined with a unique suffix.
          The provided value has the same validation rules as the Name field,
          and may be truncated by the length of the suffix required to make the value
          unique on the server.

          If this field is specified and the generated name exists, the server will
          NOT return a 409 - instead, it will either return 201 Created or 500 with Reason
          ServerTimeout indicating a unique name could not be found in the time allotted, and the client
          should retry (optionally after the time indicated in the Retry-After header).

          Applied only if Name is not specified.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#idempotency
          +optional
        type: string
      generation:
        description: |-
          A sequence number representing a specific generation of the desired state.
          Populated by the system. Read-only.
          +optional
        type: integer
      labels:
        additionalProperties:
          type: string
        description: |-
          Map of string keys and values that can be used to organize and categorize
          (scope and select) objects. May match selectors of replication controllers
          and services.
          More info: http://kubernetes.io/docs/user-guide/labels
          +optional
        type: object
      managedFields:
        description: |-
          ManagedFields maps workflow-id and version to the set of fields
          that are managed by that workflow. This is mostly for internal
          housekeeping, and users typically shouldn't need to set or
          understand this field. A workflow can be the user's name, a
          controller's name, or the name of a specific apply path like
          "ci-cd". The set of fields is always in the version that the
          workflow used when modifying the object.

          +optional
        items:
          $ref: '#/definitions/v1.ManagedFieldsEntry'
        type: array
      name:
        description: |-
          Name must be unique within a namespace. Is required when creating resources, although
          some resources may allow a client to request the generation of an appropriate name
          automatically. Name is primarily intended for creation idempotence and configuration
          definition.
          Cannot be updated.
          More info: http://kubernetes.io/docs/user-guide/identifiers#names
          +optional
        type: string
      namespace:
        description: |-
          Namespace defines the space within which each name must be unique. An empty namespace is
          equivalent to the "default" namespace, but "default" is the canonical representation.
          Not all objects are required to be scoped to a namespace - the value of this field for
          those objects will be empty.

          Must be a DNS_LABEL.
          Cannot be updated.
          More info: http://kubernetes.io/docs/user-guide/namespaces
          +optional
        type: string
      ownerReferences:
        description: |-
          List of objects depended by this object. If ALL objects in the list have
          been deleted, this object will be garbage collected. If this object is managed by a controller,
          then an entry in this list will point to this controller, with the controller field set to true.
          There cannot be more than one managing controller.
          +optional
          +patchMergeKey=uid
          +patchStrategy=merge
        items:
          $ref: '#/definitions/v1.OwnerReference'
        type: array
      resourceVersion:
        description: |-
          An opaque value that represents the internal version of this object that can
          be used by clients to determine when objects have changed. May be used for optimistic
          concurrency, change detection, and the watch operation on a resource or set of resources.
          Clients must treat these values as opaque and passed unmodified back to the server.
          They may only be valid for a particular resource or set of resources.

          Populated by the system.
          Read-only.
          Value must be treated as opaque by clients and .
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#concurrency-control-and-consistency
          +optional
        type: string
      selfLink:
        description: |-
          SelfLink is a URL representing this object.
          Populated by the system.
          Read-only.

          DEPRECATED
          Kubernetes will stop propagating this field in 1.20 release and the field is planned
          to be removed in 1.21 release.
          +optional
        type: string
      uid:
        description: |-
          UID is the unique in time and space value for this object. It is typically generated by
          the server on successful creation of a resource and is not allowed to change on PUT
          operations.

          Populated by the system.
          Read-only.
          More info: http://kubernetes.io/docs/user-guide/identifiers#uids
          +optional
        type: string
    type: object
...
  v1alpha1.LocalVolumeConvertStatus:
    properties:
      message:
        type: string
      state:
        $ref: '#/definitions/v1alpha1.State'
    type: object
...
  v1alpha1.LocalVolumeSpec:
    properties:
      accessibility:
        allOf:
        - $ref: '#/definitions/v1alpha1.AccessibilityTopology'
        description: Accessibility is the topology requirement of the volume. It describes
          how to locate and distribute the volume replicas
      config:
        allOf:
        - $ref: '#/definitions/v1alpha1.VolumeConfig'
        description: |-
          Config is the configration for the volume replicas
          It will be managed by the controller, and watched by all the nodes
          Important: node will manage volume replica according this config
      convertible:
        description: |-
          Convertible is to indicate if the non-HA volume can be transitted to HA volume or not
          +kubebuilder:default:=false
        type: boolean
      delete:
        description: |-
          Delete is to indicate where the replica should be deleted or not.
          It's different from the regular resource delete interface in Kubernetes.
          The purpose is to protect it from any mistakes
          +kubebuilder:default:=false
        type: boolean
      poolName:
        description: PoolName is the name of the storage pool, e.g. LocalStorage_PoolHDD,
          LocalStorage_PoolSSD, etc..
        type: string
      pvcName:
        description: PersistentVolumeClaimName is the name of the associated PVC
        type: string
      pvcNamespace:
        description: PersistentVolumeClaimNamespace is the namespace of the associated
          PVC
        type: string
      replicaNumber:
        description: |-
          replica number: 1 - non-HA, 2 - HA, 4 - migration (temp)
          +kubebuilder:validation:Minimum:=1
          +kubebuilder:validation:Maximum:=4
        type: integer
      requiredCapacityBytes:
        description: +kubebuilder:validation:Minimum:=4194304
        type: integer
      volumeQoS:
        allOf:
        - $ref: '#/definitions/v1alpha1.VolumeQoS'
        description: VolumeQoS is the QoS of the volume
      volumegroup:
        description: VolumeGroup is the group name of the local volumes. It is designed
          for the scheduling and allocating.
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /cluster/volumes:
    get:
      consumes:
      - application/json
      description: list Volume
      parameters:
      - description: page
        in: query
        name: page
        required: true
        type: integer
      - description: pageSize
        in: query
        name: pageSize
        required: true
        type: integer
      - description: volumeName
        in: query
        name: volumeName
        type: string
      - description: state
        in: query
        name: state
        type: string
      - description: fuzzy
        in: query
        name: fuzzy
        type: boolean
      - description: sort
        in: query
        name: sort
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.VolumeList'
      summary: 摘要 获取数据卷列表信息 数据卷状态枚举 （ToBeMounted、Created、Creating、Ready、NotReady、ToBeDeleted、Deleted）
      tags:
      - Volume
  /cluster/volumes/{volumeName}:
    get:
      consumes:
      - application/json
      description: get Volume angel1
      parameters:
      - description: volumeName
        in: path
        name: volumeName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Volume'
      summary: 摘要 获取指定数据卷基本信息
      tags:
      - Volume
  /cluster/volumes/{volumeName}/convert:
    get:
      consumes:
      - application/json
      description: get GetVolumeConvertOperation 状态枚举 （Submitted、InProgress、Completed、ToBeAborted、Aborted）
      parameters:
      - description: volumeName
        in: path
        name: volumeName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.VolumeConvertOperation'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.RspFailBody'
      summary: 摘要 获取指定数据卷转换操作
      tags:
      - Volume
    post:
      consumes:
      - application/json
      description: post VolumeConvertOperation
      parameters:
      - description: volumeName
        in: path
        name: volumeName
        required: true
        type: string
      - description: reqBody
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.VolumeConvertReqBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.VolumeConvertRspBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.RspFailBody'
      summary: 摘要 指定数据卷转换操作
      tags:
      - Volume
  /cluster/volumes/{volumeName}/migrate:
    get:
      consumes:
      - application/json
      description: get GetVolumeMigrateOperation 状态枚举 （Submitted、AddReplica、SyncReplica、PruneReplica、InProgress、Completed、ToBeAborted、Cancelled、Aborted、Failed）
      parameters:
      - description: volumeName
        in: path
        name: volumeName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.VolumeMigrateOperation'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.RspFailBody'
      summary: 摘要 获取指定数据卷迁移操作
      tags:
      - Volume
    post:
      consumes:
      - application/json
      description: post VolumeMigrateOperation body i.g. body { SrcNode string ,SelectedNode
        string}
      parameters:
      - description: volumeName
        in: path
        name: volumeName
        required: true
        type: string
      - description: reqBody
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.VolumeMigrateReqBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.VolumeMigrateRspBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.RspFailBody'
      summary: 摘要 指定数据卷迁移操作
      tags:
      - Volume
  /cluster/volumes/{volumeName}/operations:
    get:
      consumes:
      - application/json
      description: get VolumeOperation
      parameters:
      - description: volumeName
        in: path
        name: volumeName
        required: true
        type: string
      - description: volumeEventName
        in: query
        name: volumeEventName
        type: string
      - description: state
        in: query
        name: state
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.VolumeOperationByVolume'
      summary: 摘要 获取指定数据卷操作记录信息 状态枚举 (Submitted、AddReplica、SyncReplica、PruneReplica、InProgress、Completed、ToBeAborted、Cancelled、Aborted、Failed)
      tags:
      - Volume
  /cluster/volumes/{volumeName}/replicas:
    get:
      consumes:
      - application/json
      description: list volumes
      parameters:
      - description: volumeName
        in: path
        name: volumeName
        required: true
        type: string
      - description: volumeReplicaName
        in: query
        name: volumeReplicaName
        type: string
      - description: state
        in: query
        name: state
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.VolumeReplicaList'
      summary: 摘要 获取指定数据卷的副本列表信息
      tags:
      - Volume
swagger: "2.0"